fix: k8s cluster group_vars relative path instead of absolute
Before we included the absolute path inside the symbolic link of
group_vars. This is not okay since we push this sym link to git,
which only works for the user who created it (since it contains
their $HOME). With this fix we use relative paths in order to
mitigate this issue.